Cleanup

+++ scripts/wandb_clean.py
@@ -1,13 +1,21 @@
+"""
+Cleans up test runs and dev runs
+"""
+
 import wandb
 from wandb.wandb_run import Run
 
 # Replace 'entity/project' with your specific entity and project name
 entity_project = "harrycblum/nano-diarization"
 
+
 api = wandb.Api()
-runs:list[Run] = api.runs(entity_project)
+runs: list[Run] = api.runs(entity_project)
 
 for run in runs:
+    if run.state == "running":
+        continue
+
     if run.name == "dev":
         run.delete()
         continue
@@ -15,7 +23,7 @@ for run in runs:
     if not len(run.summary.keys()):
         run.delete()
         continue
-    
+
     if "_wandb" not in run.summary:
         run.delete()
         continue
@@ -23,7 +31,7 @@ for run in runs:
     info = run.summary
     # print(run.summary)
     runtime = run.summary.get("_wandb").get("runtime")
-    
+
     if "train/loss" not in info:
         run.delete()
         continue
@@ -31,11 +39,8 @@ for run in runs:
     if not runtime or runtime < 60:
         run.delete()
         continue
-    
+
     if "_step" in info and info["_step"] < 500:
         print(info["_step"])
         run.delete()
         continue
-
-    
-
